test(rng): add fixture-based parity suite and fix two bit-level bugs
Loads the binary fixtures emitted by fixtures-gen and replays every
input through the Rust JavaRng / Xoroshiro / mc_seed APIs, comparing
outputs bit-exactly (f32/f64 via to_bits). The suite catches two bugs
the unit tests missed:

- JavaRng::next_long sign-extends each next(32) result through i64
  instead of zero-extending through u32; cubiomes' uint64_t cast on a
  signed int is sign-extending under C99, so the previous Rust path
  produced different upper bits whenever next(32) returned a negative
  value.
- mc_first_int now performs `(s >> 24) % m` in i64 before truncating
  to i32. The old order truncated first, which dropped the upper 32
  bits of the shifted seed and yielded a different remainder when the
  shifted value did not fit in i32.

All 10 000 records per fixture now match cubiomes byte-for-byte.

+++ core/src/rng/mc_seed.rs
@@ -29,11 +29,14 @@ pub const fn mc_step_seed(s: u64, salt: u64) -> u64 {
 /// First PRNG integer in `[0, m)` derived from a chunk seed.
 ///
 /// Mirrors `mcFirstInt`: the seed is treated as `i64`, arithmetic-shifted
-/// right by 24, taken modulo `m`, and adjusted into the positive range.
+/// right by 24, taken modulo `m` *in 64-bit precision*, then truncated to
+/// `i32` and adjusted into the positive range. The order matters — taking
+/// `as i32` before `% m` would drop the upper 32 bits and yield a
+/// different remainder whenever `s >> 24` is outside `i32` range.
 #[inline]
 #[must_use]
 pub const fn mc_first_int(s: u64, m: i32) -> i32 {
-    let r = ((s as i64) >> 24) as i32 % m;
+    let r = (((s as i64) >> 24) % (m as i64)) as i32;
     if r < 0 { r + m } else { r }
 }
 
@@ -42,7 +45,7 @@ pub const fn mc_first_int(s: u64, m: i32) -> i32 {
 #[inline]
 #[must_use]
 pub const fn mc_first_is_zero(s: u64, m: i32) -> bool {
-    (((s as i64) >> 24) as i32 % m) == 0
+    (((s as i64) >> 24) % (m as i64)) == 0
 }
 
 /// Combine the per-layer start seed with chunk coordinates.
